requirements.txt
amqp==2.2.2
appnope==0.1.0
billiard==3.5.0.3
celery==4.1.0
decorator==4.1.2
Django==1.11.7
django-celery==3.2.1
django-mailing==0.1.3
djangorestframework==3.7.3
gunicorn==19.7.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
kombu==4.1.0
parso==0.1.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
whitenoise==3.3.1
git+https://github.com/razortheory/drf-secure-token